refactor: separate concerns in TagManager — hooks, table, drawer
- Extract useTagFilter hook for filter/sort logic
- Extract getImpactedItemsCount to TagManager.utils.js
- Move rename/delete state, mutations, dialogs into TagManagerTable
- Move editNode/deleteNode state, mutations, dialogs into TagManagerDrawer
- Move table pagination into TagManagerTable; drawer pagination into TagManagerDrawer
- TagManager now only orchestrates data fetching and selectedTag state
